Fixing comments from code review
