+++ wodles/aws/tests/test_aws.py
@@ -20,10 +20,8 @@ sys.path.append(os.path.join(os.path.dirname(os.path.realpath(__file__)), '..'))
 import aws_s3
 
 test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data')
-wazuh_control_info = 'WAZUH_VERSION="TEST_VERSION"\n\
-                      WAZUH_REVISION="TEST_REVISION"\n\
-                      WAZUH_TYPE="TEST_TYPE"\n'
 wazuh_installation_path = '/var/ossec'
+wazuh_version = 'WAZUH_VERSION'
 
 def get_fake_s3_db(sql_file):
 
@@ -52,8 +50,8 @@ def test_metadata_version_buckets(mocked_db, class_):
     """
     with patch(f'aws_s3.{class_.__name__}.get_client'), \
         patch(f'aws_s3.{class_.__name__}.get_sts_client'), \
-        patch(f'aws_s3.call_wazuh_control', return_value=wazuh_control_info), \
-        patch(f'aws_s3.get_wazuh_path', return_value=wazuh_installation_path):
+        patch(f'common.find_wazuh_path', return_value=wazuh_installation_path), \
+        patch(f'common.get_wazuh_version', return_value=wazuh_version):
         ins = class_(**{'reparse': False, 'access_key': None, 'secret_key': None,
                         'profile': None, 'iam_role_arn': None, 'bucket': 'test',
                         'only_logs_after': '19700101', 'skip_on_error': True,
@@ -77,8 +75,8 @@ def test_metadata_version_services(mocked_db, class_):
     """
     with patch(f'aws_s3.{class_.__name__}.get_client'), \
         patch(f'aws_s3.{class_.__name__}.get_sts_client'), \
-        patch(f'aws_s3.call_wazuh_control', return_value=wazuh_control_info), \
-        patch(f'aws_s3.get_wazuh_path', return_value=wazuh_installation_path):
+        patch(f'common.find_wazuh_path', return_value=wazuh_installation_path), \
+        patch(f'common.get_wazuh_version', return_value=wazuh_version):
         ins = class_(**{'reparse': False, 'access_key': None, 'secret_key': None,
                         'aws_profile': None, 'iam_role_arn': None,
                         'only_logs_after': '19700101', 'region': None})
@@ -103,8 +101,8 @@ def test_db_maintenance(class_, sql_file, db_name):
     with patch(f'aws_s3.{class_.__name__}.get_client'), \
         patch(f'aws_s3.{class_.__name__}.get_sts_client'), \
         patch('sqlite3.connect', side_effect=get_fake_s3_db(sql_file)), \
-        patch(f'aws_s3.call_wazuh_control', return_value=wazuh_control_info), \
-        patch(f'aws_s3.get_wazuh_path', return_value=wazuh_installation_path):
+        patch(f'common.find_wazuh_path', return_value=wazuh_installation_path), \
+        patch(f'common.get_wazuh_version', return_value=wazuh_version):
         ins = class_(**{'reparse': False, 'access_key': None, 'secret_key': None,
                         'profile': None, 'iam_role_arn': None, 'bucket': 'test-bucket',
                         'only_logs_after': '19700101', 'skip_on_error': True,
