+++ wodles/gcloud/integration.py
@@ -13,6 +13,7 @@ import socket
 
 from google.api_core import exceptions as google_exceptions
 from google.cloud import pubsub_v1 as pubsub
+from wazuh.core import common
 
 import tools
 
@@ -33,9 +34,9 @@ class WazuhGCloudSubscriber:
         :params subscription_id: Subscription ID
         """
         # get Wazuh paths
-        self.wazuh_path = tools.get_wazuh_path()
+        self.wazuh_path = common.find_wazuh_path()
         self.wazuh_queue = tools.get_wazuh_queue()
-        self.wazuh_version = tools.get_wazuh_version()
+        self.wazuh_version = common.get_wazuh_version()
         # get subscriber
         self.subscriber = self.get_subscriber_client(credentials_file).api
         self.subscription_path = self.get_subscription_path(project,
